safari-534.53-branch/LayoutTests/fast/dom/Range/select-node-different-document.html

<!DOCTYPE html>
<html>
<head>
<link rel="stylesheet" href="../../js/resources/js-test-style.css">
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<body>
<iframe src="about:blank"></iframe>
<script>
description("Test whether Range.selectNode and Range.selectNodeContents work across documents");

var otherDocument = document.implementation.createDocument(null, "doc", null);
var elem = otherDocument.createElement("elem");
otherDocument.documentElement.appendChild(elem);

var selectNodeRange = document.createRange();
selectNodeRange.selectNode(elem);

shouldBeTrue("selectNodeRange.startContainer === otherDocument.documentElement");
shouldBeTrue("selectNodeRange.endContainer === otherDocument.documentElement");
shouldBeTrue("selectNodeRange.startContainer.ownerDocument === otherDocument");
shouldBeTrue("selectNodeRange.endContainer.ownerDocument === otherDocument");
shouldBeTrue("selectNodeRange.cloneContents().ownerDocument === otherDocument");

var selectNodeContentsRange = document.createRange();
selectNodeContentsRange.selectNodeContents(elem);

shouldBeTrue("selectNodeRange.startContainer === otherDocument.documentElement");
shouldBeTrue("selectNodeRange.endContainer === otherDocument.documentElement");
shouldBeTrue("selectNodeContentsRange.startContainer.ownerDocument === otherDocument");
shouldBeTrue("selectNodeContentsRange.endContainer.ownerDocument === otherDocument");
shouldBeTrue("selectNodeRange.cloneContents().ownerDocument === otherDocument");

var successfullyParsed = true;
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
