third_party/blink/web_tests/fast/dom/Range/select-node-different-document-expected.txt

Test whether Range.selectNode and Range.selectNodeContents work across documents

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS selectNodeRange.startContainer === otherDocument.documentElement is true
PASS selectNodeRange.endContainer === otherDocument.documentElement is true
PASS selectNodeRange.startContainer.ownerDocument === otherDocument is true
PASS selectNodeRange.endContainer.ownerDocument === otherDocument is true
PASS selectNodeRange.cloneContents().ownerDocument === otherDocument is true
PASS selectNodeRange.startContainer === otherDocument.documentElement is true
PASS selectNodeRange.endContainer === otherDocument.documentElement is true
PASS selectNodeContentsRange.startContainer.ownerDocument === otherDocument is true
PASS selectNodeContentsRange.endContainer.ownerDocument === otherDocument is true
PASS selectNodeRange.cloneContents().ownerDocument === otherDocument is true
PASS successfullyParsed is true

TEST COMPLETE
